[#162457306] - Organize new react form files
