+++ app/javascript/components/applications/application_form/PaperApplicationForm.js
@@ -156,15 +156,14 @@ class PaperApplicationForm extends React.Component {
           initialValues={application}
           validate={ values => {
             const errors = {applicant: {date_of_birth: {}}}
-            if (values.applicant.date_of_birth) {
+            if (values.applicant && values.applicant.date_of_birth) {
               let DOB = [values.applicant.date_of_birth.year, values.applicant.date_of_birth.month, values.applicant.date_of_birth.day]
+              console.log(errors)
               errors.applicant.date_of_birth.all = validate.any(
                 validate.isPresent('Please enter a Date of Birth'),
                 validate.isValidDate('Please enter a valid Date of Birth'),
                 validate.isOldEnough('The primary applicant must be 18 years of age or older')
               )(DOB)
-            } else {
-              errors.applicant.date_of_birth.all = 'Shouldnt be empty!'
             }
             return errors
           }}
