+++ app/javascript/utils/form/final_form/MultiDateField.js
@@ -1,20 +1,20 @@
 import React from 'react'
 import { BlockNote } from '~/utils/form/Field'
 import { maxLengthMap } from '~/utils/formUtils'
-import { FieldWrapper } from '~/utils/form/Field2'
+import { FieldWrapper } from '~/utils/form/final_form/Field'
+import reactFormUtils from '~/utils/reactFormUtils'
 
-export const MultiDateField2 = ({ form, fieldName, label, blockNote }) => {
-  const errors = form.getState().errors.applicant && form.getState().errors.applicant.date_of_birth.all
+
+export const MultiDateField = ({ form, fieldName, label, blockNote }) => {
+  const error = form.getState && form.getState().errors.applicant && form.getState().errors.applicant.date_of_birth.all
   return (
-    // <div className={classNames('form-group', className)}>
-    <div>
+    <div className={(error && 'error') || ''}>
       <label className='form-label' htmlFor={fieldName}>
         {`${label} `}
         {blockNote && <BlockNote value={blockNote} />}
       </label>
       <div className='form-group-month'>
         <FieldWrapper
-          // className={classNames(className)}
           type='text'
           fieldName={`${fieldName}.month`}
           placeholder='MM'
@@ -23,7 +23,6 @@ export const MultiDateField2 = ({ form, fieldName, label, blockNote }) => {
       </div>
       <div className='form-group-day'>
         <FieldWrapper
-          // className={classNames(className)}
           type='text'
           fieldName={`${fieldName}.day`}
           placeholder='DD'
@@ -32,7 +31,6 @@ export const MultiDateField2 = ({ form, fieldName, label, blockNote }) => {
       </div>
       <div className='form-group-year'>
         <FieldWrapper
-          // className={classNames(className)}
           type='text'
           fieldName={`${fieldName}.year`}
           placeholder='YYYY'
@@ -40,14 +38,8 @@ export const MultiDateField2 = ({ form, fieldName, label, blockNote }) => {
           maxLength={maxLengthMap['year']} />
       </div>
       <div className='d-inline-block'>
-        { errors && <span className='error'>{errors}</span> }
+        { error && <span className='small error'>{error}</span> }
       </div>
     </div>
   )
 }
-
-// 1. Display error message in multidatefield2 section
-// 2. Display error class on the date fields
-// 3. Validate with all three validations
-
-        // {/* <FormError formApi={formApi} label={label} fieldName={fieldName} errorMessage={errorMessage} /> */}
